later-backend/requirements.txt
alembic==1.13.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
supabase==2.0.3